initial app update for first trials
