+++ app/insert_table.py
@@ -1,4 +1,4 @@
-from db_connection import get_connection
+from app.db_connection import get_connection
 
 # Get the database connection
 mydb = get_connection()
