--- docker/daemon.json
@@ -1,14 +1,0 @@
-{
-  "log-driver": "syslog",
-  "log-opts": {
-    "syslog-address": "tcp://192.168.1.10:514",
-    "syslog-facility": "daemon",
-    "syslog-format": "rfc5424",
-    "tag": "{{.Name}}({.ID}})"
-  }
-}
-  
-
-
-"syslog-format": "rfc5424micro",
-"tag": "docker-{{.Name}}-{{.ID}}"
