--- docker/ssl.conf
@@ -1,13 +1,0 @@
-sudo docker exec -it test_db bash
-
-vim /etc/mysql/my.cnf
-
-ssl=0
-skip_ssl
-
-
-ALTER USER 'ken'@'%' REQUIRE NONE;
-FLUSH PRIVILEGES;
-
-
-Password4321reqW
