+++ web/default.conf
@@ -1,0 +1,9 @@
+server {
+    listen 80;
+
+    location / {
+        proxy_pass http://app:5000;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+    }
+}
